refactor(mem): move src/lib/sync → src/lib/memory/sync
Sync is structurally part of the unified memory subsystem now — its
hot path writes through upsertByKeys into mem_records, and the next
slice kills the SQLite fallback entirely. The folder relocation makes
that ownership explicit.

Pure rename. Internal relative imports rewritten:
  ../memory/X → ../X   (now siblings inside memory/)
  ../X        → ../../X (lib/ peers one level further out)

External importers fixed:
  src/index.ts                — registerSyncCommands path
  src/commands/mem/migrate.ts — readProfile, openDatabase, etc.

Zero behavior changes. All 61 tests green; `sync list`, `sync test`
still function.
