+++ src/lib/memory/sync/extract.test.ts
