+++ src/lib/memory/sync/index.ts
@@ -1,7 +1,7 @@
 import type { Command } from 'commander';
-import * as output from '../output.js';
-import { OneApi } from '../api.js';
-import { getApiKey, getAccessControlFromAllSources } from '../config.js';
+import * as output from '../../output.js';
+import { OneApi } from '../../api.js';
+import { getApiKey, getAccessControlFromAllSources } from '../../config.js';
 import { discoverModels } from './models.js';
 import { readProfile, writeProfile, writeDraftProfile, listProfiles, generateTemplate } from './profile.js';
 import { syncModel } from './runner.js';
@@ -14,7 +14,7 @@ import { readSyncState, removeModelState, getModelState } from './state.js';
 import { executeQuery, executeRawSql } from './query.js';
 import { searchSyncedData } from './search.js';
 import { extractSearchableFromPaths, getSearchablePaths } from './mem-writer.js';
-import { defaultSearchableText } from '../memory/embedding.js';
+import { defaultSearchableText } from '../embedding.js';
 import { openDatabase, getDatabaseSize, dropTable, deleteDatabase, countRecords, listTables, deleteRecords, rebuildFtsIndex, tableExists, sanitizeTableName } from './db.js';
 import type { SyncProfile, SyncRunOptions, SyncQueryOptions } from './types.js';
 import { isSqliteAvailable, loadSqlite } from './sqlite-loader.js';
