+++ src/lib/memory/sync/lock.ts
